[00013] Rename Settings folder and views to Setup
Rename Apps/Settings/ to Apps/Setup/, rename all *SettingsView classes
to *SetupView, and update namespaces from Ivy.Tendril.Apps.Settings
to Ivy.Tendril.Apps.Setup to match the existing SetupApp naming.
