+++ src/tendril/Ivy.Tendril/Apps/Setup/Dialogs/EditPromptwareDialog.cs
@@ -1,6 +1,6 @@
 using Ivy.Tendril.Services;
 
-namespace Ivy.Tendril.Apps.Settings.Dialogs;
+namespace Ivy.Tendril.Apps.Setup.Dialogs;
 
 public class EditPromptwareDialog(
     IState<string?> editKey,
