+++ src/tendril/Ivy.Tendril/Apps/Setup/Dialogs/EditVerificationDialog.cs
@@ -1,16 +1,16 @@
 using Ivy.Tendril.Services;
 
-namespace Ivy.Tendril.Apps.Settings.Dialogs;
+namespace Ivy.Tendril.Apps.Setup.Dialogs;
 
-public class EditLevelDialog(
+public class EditVerificationDialog(
     IState<int?> editIndex,
-    List<LevelConfig> levels,
+    List<VerificationConfig> verifications,
     IConfigService config,
     IClientProvider client,
     RefreshToken refreshToken) : ViewBase
 {
     private readonly IState<int?> _editIndex = editIndex;
-    private readonly List<LevelConfig> _levels = levels;
+    private readonly List<VerificationConfig> _verifications = verifications;
     private readonly IConfigService _config = config;
     private readonly IClientProvider _client = client;
     private readonly RefreshToken _refreshToken = refreshToken;
@@ -18,34 +18,33 @@ public class EditLevelDialog(
     public override object? Build()
     {
         var editName = UseState("");
-        var editBadge = UseState("Outline");
+        var editPrompt = UseState("");
 
         UseEffect(() =>
         {
             if (_editIndex.Value == null)
             {
                 editName.Set("");
-                editBadge.Set("Outline");
+                editPrompt.Set("");
             }
             else if (_editIndex.Value >= 0)
             {
-                editName.Set(_levels[_editIndex.Value.Value].Name);
-                editBadge.Set(_levels[_editIndex.Value.Value].Badge);
+                editName.Set(_verifications[_editIndex.Value.Value].Name);
+                editPrompt.Set(_verifications[_editIndex.Value.Value].Prompt);
             }
         }, _editIndex);
 
         if (_editIndex.Value == -1) return null;
 
-        var badgeOptions = Enum.GetNames<BadgeVariant>().ToList();
         var isNew = _editIndex.Value == null;
 
         return new Dialog(
-            _ => { _editIndex.Set(-1); },
-            new DialogHeader(isNew ? "Add Level" : "Edit Level"),
+            _ => _editIndex.Set(-1),
+            new DialogHeader(isNew ? "Add Verification" : "Edit Verification"),
             new DialogBody(
                 Layout.Vertical().Gap(2)
-                | editName.ToTextInput("Level name...").WithField().Label("Name")
-                | editBadge.ToSelectInput(badgeOptions).WithField().Label("Badge Variant")
+                | editName.ToTextInput("Verification name...").WithField().Label("Name")
+                | editPrompt.ToTextareaInput("Verification prompt...").Rows(8).WithField().Label("Prompt")
             ),
             new DialogFooter(
                 new Button("Cancel").Outline().OnClick(() => _editIndex.Set(-1)),
@@ -54,21 +53,24 @@ public class EditLevelDialog(
                     if (string.IsNullOrWhiteSpace(editName.Value)) return;
                     if (isNew)
                     {
-                        _levels.Add(new LevelConfig { Name = editName.Value, Badge = editBadge.Value });
+                        _verifications.Add(new VerificationConfig
+                        {
+                            Name = editName.Value,
+                            Prompt = editPrompt.Value
+                        });
                     }
                     else
                     {
-                        var level = _levels[_editIndex.Value!.Value];
-                        level.Name = editName.Value;
-                        level.Badge = editBadge.Value;
+                        _verifications[_editIndex.Value!.Value].Name = editName.Value;
+                        _verifications[_editIndex.Value!.Value].Prompt = editPrompt.Value;
                     }
 
                     _config.SaveSettings();
                     _editIndex.Set(-1);
                     _refreshToken.Refresh();
-                    _client.Toast("Level saved", "Saved");
+                    _client.Toast("Verification saved", "Saved");
                 })
             )
-        ).Width(Size.Rem(25));
+        ).Width(Size.Rem(35));
     }
 }
