+++ src/tendril/Ivy.Tendril/Apps/Setup/LevelsSetupView.cs
@@ -1,9 +1,9 @@
-using Ivy.Tendril.Apps.Settings.Dialogs;
+using Ivy.Tendril.Apps.Setup.Dialogs;
 using Ivy.Tendril.Services;
 
-namespace Ivy.Tendril.Apps.Settings;
+namespace Ivy.Tendril.Apps.Setup;
 
-public class LevelsSettingsView : ViewBase
+public class LevelsSetupView : ViewBase
 {
     public override object Build()
     {
